moved get_homeserver from internal to auth
