+++ pubky/src/native/internal/pkarr.rs
@@ -111,31 +111,6 @@ impl Client {
             })
     }
 
-    /// Get the homeserver for a given Pubky public key.
-    /// Looks up the pkarr packet for the given public key and returns the content of the first `_pubky` SVCB record.
-    pub async fn get_homeserver(&self, pubky: &PublicKey) -> Option<String> {
-        let packet = self.pkarr.resolve_most_recent(pubky).await?;
-
-        // Check for the `_pubky` SVCB record.
-        let name = format!("_pubky.{}", pubky.to_z32());
-        let maching_names = packet.resource_records(name.as_str()).collect::<Vec<_>>();
-
-        let pubky_records = maching_names
-            .into_iter()
-            .map(|r| r.rdata.clone())
-            .filter(|r| matches!(r, RData::HTTPS(_)))
-            .collect::<Vec<_>>();
-
-        if pubky_records.is_empty() {
-            return None;
-        }
-
-        let record = pubky_records.first().unwrap();
-        if let RData::HTTPS(svc) = record {
-            return Some(svc.target.to_string());
-        }
-        None
-    }
 }
 
 #[cfg(test)]
@@ -163,26 +138,4 @@ mod tests {
         Ok(())
     }
 
-    #[tokio::test]
-    async fn test_get_homeserver() {
-        let dht = mainline::Testnet::new(3).unwrap();
-        let client = Client::builder()
-            .pkarr(|builder| builder.bootstrap(&dht.bootstrap))
-            .build()
-            .unwrap();
-        let keypair = Keypair::random();
-        let pubky = keypair.public_key();
-
-        let homeserver_key = Keypair::random().public_key().to_z32();
-        client
-            .publish_homeserver(
-                &keypair,
-                Some(homeserver_key.as_str()),
-                PublishStrategy::Force,
-            )
-            .await
-            .unwrap();
-        let homeserver = client.get_homeserver(&pubky).await;
-        assert_eq!(homeserver, Some(homeserver_key));
-    }
 }
